Update GLCashFlowsIndirect.php
Files changed in commit:
GLCashFlowsIndirect.php
On branch develop

Fix the accumulated total reset of $BudgetSection and $BudgetTotal. Fix width of the multicolumn elements.
Extracts the sentences that are different between the "ShowBudget=ON" and the "ShowBudget=OFF" sections to help the code maintenance. Also, it is a preparation for a later reduction of the code without adding if() sentences