Delete file?
For whatever reason, after cloning the repo, this file showed a diff? Attempts to replace locally did not change anything, so I downloaded the current copy from GitHub.

When pasting the file locally (a forced attempt to overwrite), there was a git complaint about a local resource already existing, though the file contents differ, the only difference in name is second P in the filename.

So, I had to locally rename the one with the lower-case p, and then the local copy cleared the resolved with git as if OK, and git marked this one as deleted.

Not sure of the situation, but that is the deal.
(this appears to be an older copy, but I'm not sure)