Git conflict with prlPayperiod.php
Somehow, there's a git resource conflict with this file. The name is off by 1 letter, case-wise, and the one with the lower-case p (in period) seems to have trouble?

The lower case p version would not discard changes, but after renaming to having a capital P, then that file resolved and git marks this one as deleted.